% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGramSchmidt.R
\name{MGramSchmidt}
\alias{MGramSchmidt}
\title{Gram Schmidt Orthogonormalization of a matrix}
\usage{
MGramSchmidt(V)
}
\arguments{
\item{V}{matrix to be orthogonormalized}
}
\value{
orthogonormalized matrix
}
\description{
Gram Schmidt Orthogonormalization of a matrix
}
\examples{
M = matrix(rnorm(20), 10, 2)
MGramSchmidt(M)
}

