% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptStiefelGBB.R
\name{OptStiefelGBB}
\alias{OptStiefelGBB}
\title{Gradient based optimization on the Stiefel manifold}
\usage{
OptStiefelGBB(fun, X, size, opts, ...)
}
\arguments{
\item{fun}{function to be minimized. It must take \code{X} as the first argument and
return a list with two named elements:
\code{f}, which is the value of the objective function and \code{G}, which is
the derivative}

\item{X}{the initial orthonormal matrix}

\item{size}{if initial value for \code{X} is not given, the number of rows and columns
can be given instead in a two-element vector. Number of columns must be less than the
number of rows}

\item{opts}{a list of options for the optimization}

\item{...}{additional argmuents to pass to \code{fun}}
}
\value{
list with \code{X}, which is the solution found, and \code{out},
  which includes supplementary material
}
\description{
Gradient based optimization on the Stiefel manifold
}
\details{
This R package ports the MATLAB code from the authors' website
  http://optman.blogs.rice.edu/
}
\examples{

# find the first eigenvector
fun <- function(X,  A) {
  G = -(A \%*\% X)
  f = 0.5 * sum(t(G) \%*\% X)
  list(f = f, G = G)
}

n = 1000; k = 1
A = matrix(rnorm(n^2), n, n)
A = t(A) \%*\% A

res = OptStiefelGBB(fun, size = c(n, k), A = A)

# objective function should be close to the sum of the first eigenvalue
eig = eigen(A)
sum(eig$values[1:k])
-2 * res$out$fval

plot(res$X[, 1], eig$vectors[, 1])

}
\references{
Zaiwen Wen and Wotao Yin. A Feasible method for Optimization with
  Orthogonality Constraints, Optimization Online, 11/2010. Also as Rice
  CAAM Tech Report TR10-26.
}

